requirements.txt
aiohttp
faiss_gpu==1.7.2
flagembedding==1.3.5
jieba
llama_index==0.12.48
llama-index-embeddings-huggingface==0.5.5
nltk
numpy==1.26.4
openai
rouge==1.0.1
torch==2.8.0
transformers==4.56.2
vllm==0.10.2
